requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.2.0
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==41.0.3
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
flatbuffers==20210226132247
fonttools==4.53.0
frozenlist==1.4.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.42.0
h11==0.14.0
h5py==3.11.0
idna==3.7
importlib-metadata==7.0.1
importlib_resources==6.4.0
ipykernel==6.29.3
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.1
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
multidict==6.0.4
munkres==1.1.4
nest_asyncio==1.6.0
numpy==1.23.4
oauthlib==3.2.2
olefile==0.47
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pickleshare==0.7.5
Pillow==8.3.2
pip==24.0
platformdirs==4.2.2
prompt-toolkit==3.0.42
protobuf==3.17.2
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyparsing==3.1.2
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
PySocks==1.7.1
python-dateutil==2.9.0
python-dotenv==0.21.0
pytz==2024.1
pywin32==306
pyzmq==26.0.3
requests==2.32.2
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.13.1
selenium==4.21.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.2
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==2.1.0
tornado==6.4
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
unicodedata2==15.1.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wheel==0.35.1
win-inet-pton==1.1.0
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.3
zipp==3.17.0
